+++ tests/TestCase/Utility/AssetsCreatorTest.php
@@ -13,18 +13,15 @@
 namespace Assets\Test\TestCase\Utility;
 
 use Assets\Utility\AssetsCreator;
+use Assets\TestSuite\TestCase;
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
-use Cake\TestSuite\TestCase;
-use Tools\ReflectionTrait;
 
 /**
  * AssetsCreatorTest class
  */
 class AssetsCreatorTest extends TestCase
 {
-    use ReflectionTrait;
-
     /**
      * Setup the test case, backup the static object values so they can be
      * restored. Specifically backs up the contents of Configure and paths in
@@ -49,11 +46,6 @@ class AssetsCreatorTest extends TestCase
         parent::tearDown();
 
         Plugin::unload('TestPlugin');
-
-        //Deletes all assets
-        foreach (glob(Configure::read(ASSETS . '.target') . DS . '*') as $file) {
-            unlink($file);
-        }
     }
 
     /**
@@ -64,12 +56,10 @@ class AssetsCreatorTest extends TestCase
     {
         $asset = new AssetsCreator('test', 'css');
 
-        $this->assertInstanceOf(ASSETS . '\Utility\AssetsCreator', $asset);
-
         $this->assertEquals('css', $this->getProperty($asset, 'type'));
 
         $paths = $this->getProperty($asset, 'paths');
-        $this->assertEquals(1, count($paths));
+        $this->assertCount(1, $paths);
         $this->assertEquals(WWW_ROOT . 'css', dirname($paths[0]));
         $this->assertEquals('test.css', basename($paths[0]));
 
@@ -117,22 +107,22 @@ class AssetsCreatorTest extends TestCase
      */
     public function testResolveAssetPath()
     {
-        $resolveAssetPathMethod = function (\Assets\Utility\AssetsCreator $assetCreatorInstance) {
+        $resolveAssetPathMethod = function (AssetsCreator $assetCreatorInstance) {
             return $this->invokeMethod($assetCreatorInstance, 'resolveAssetPath');
         };
 
-        $result = $resolveAssetPathMethod((new AssetsCreator('test', 'css')));
-        $this->assertEquals(Configure::read(ASSETS . '.target') . DS . sprintf('%s.%s', md5(serialize([[
+        $expected = Configure::read(ASSETS . '.target') . DS . sprintf('%s.%s', md5(serialize([[
             $file = WWW_ROOT . 'css' . DS . 'test.css',
             filemtime($file),
-        ]])), 'css'), $result);
+        ]])), 'css');
+        $this->assertEquals($expected, $resolveAssetPathMethod((new AssetsCreator('test', 'css'))));
 
         //From plugin
-        $result = $resolveAssetPathMethod((new AssetsCreator('TestPlugin.test', 'css')));
-        $this->assertEquals(Configure::read(ASSETS . '.target') . DS . sprintf('%s.%s', md5(serialize([[
+        $expected = Configure::read(ASSETS . '.target') . DS . sprintf('%s.%s', md5(serialize([[
             $file = Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'test.css',
             filemtime($file),
-        ]])), 'css'), $result);
+        ]])), 'css');
+        $this->assertEquals($expected, $resolveAssetPathMethod((new AssetsCreator('TestPlugin.test', 'css'))));
     }
 
     /**
@@ -141,28 +131,33 @@ class AssetsCreatorTest extends TestCase
      */
     public function testResolveFilePaths()
     {
-        $pathsProperty = function (\Assets\Utility\AssetsCreator $assetCreatorInstance) {
+        $pathsProperty = function (AssetsCreator $assetCreatorInstance) {
             return $this->getProperty($assetCreatorInstance, 'paths');
         };
 
-        foreach (['test', 'test.css', '/css/test', '/css/test.css'] as $path) {
-            $result = $pathsProperty(new AssetsCreator($path, 'css'));
-            $this->assertEquals([WWW_ROOT . 'css' . DS . 'test.css'], $result);
+        $expected = [WWW_ROOT . 'css' . DS . 'test.css'];
+        foreach ([
+            'test',
+            'test.css',
+            '/css/test',
+            '/css/test.css',
+        ] as $path) {
+            $this->assertEquals($expected, $pathsProperty(new AssetsCreator($path, 'css')));
         }
 
-        $result = $pathsProperty(new AssetsCreator('subdir/test', 'css'));
-        $this->assertEquals([WWW_ROOT . 'css' . DS . 'subdir' . DS . 'test.css'], $result);
-
-        $result = $pathsProperty(new AssetsCreator('/othercssdir/test', 'css'));
-        $this->assertEquals([WWW_ROOT . 'othercssdir' . DS . 'test.css'], $result);
+        foreach ([
+            'subdir/test' => [WWW_ROOT . 'css' . DS . 'subdir' . DS . 'test.css'],
+            '/othercssdir/test' => [WWW_ROOT . 'othercssdir' . DS . 'test.css'],
+        ] as $path => $expected) {
+            $this->assertEquals($expected, $pathsProperty(new AssetsCreator($path, 'css')));
+        }
 
         //Tests array
-        $asset = new AssetsCreator([
+        $result = $pathsProperty(new AssetsCreator([
             'test',
             'subdir/test',
             '/othercssdir/test',
-        ], 'css');
-        $result = $pathsProperty($asset);
+        ], 'css'));
         $this->assertEquals([
             WWW_ROOT . 'css' . DS . 'test.css',
             WWW_ROOT . 'css' . DS . 'subdir' . DS . 'test.css',
@@ -170,43 +165,34 @@ class AssetsCreatorTest extends TestCase
         ], $result);
 
         //Tests plugins
+        $expected = [Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'test.css'];
         foreach ([
             'TestPlugin.test',
             'TestPlugin.test.css',
             'TestPlugin./css/test',
             'TestPlugin./css/test.css',
         ] as $path) {
-            $result = $pathsProperty(new AssetsCreator($path, 'css'));
-            $this->assertEquals(
-                [Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'test.css'],
-                $result
-            );
+            $this->assertEquals($expected, $pathsProperty(new AssetsCreator($path, 'css')));
         }
 
-        $result = $pathsProperty(new AssetsCreator('TestPlugin.subdir/test', 'css'));
-        $this->assertEquals(
-            [Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'subdir' . DS . 'test.css'],
-            $result
-        );
+        $expected = [Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'subdir' . DS . 'test.css'];
+        $this->assertEquals($expected, $pathsProperty(new AssetsCreator('TestPlugin.subdir/test', 'css')));
 
-        $result = $pathsProperty(new AssetsCreator('TestPlugin./othercssdir/test', 'css'));
-        $this->assertEquals(
-            [Plugin::path('TestPlugin') . 'webroot' . DS . 'othercssdir' . DS . 'test.css'],
-            $result
-        );
+        $expected = [Plugin::path('TestPlugin') . 'webroot' . DS . 'othercssdir' . DS . 'test.css'];
+        $this->assertEquals($expected, $pathsProperty(new AssetsCreator('TestPlugin./othercssdir/test', 'css')));
 
         //Tests array
-        $asset = new AssetsCreator([
-            'TestPlugin.test',
-            'TestPlugin.subdir/test',
-            'TestPlugin./othercssdir/test'
-        ], 'css');
-        $result = $pathsProperty($asset);
-        $this->assertEquals([
+        $expected = [
             Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'test.css',
             Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'subdir' . DS . 'test.css',
             Plugin::path('TestPlugin') . 'webroot' . DS . 'othercssdir' . DS . 'test.css',
-        ], $result);
+        ];
+        $result = $pathsProperty(new AssetsCreator([
+            'TestPlugin.test',
+            'TestPlugin.subdir/test',
+            'TestPlugin./othercssdir/test'
+        ], 'css'));
+        $this->assertEquals($expected, $result);
     }
 
     /**
@@ -216,7 +202,7 @@ class AssetsCreatorTest extends TestCase
     public function testCreateWithCss()
     {
         $result = (new AssetsCreator('test', 'css'))->create();
-        $this->assertRegExp('/^[a-z0-9]+$/', $result);
+        $this->assertRegExp('/^[\w\d]+$/', $result);
 
         $file = Configure::read(ASSETS . '.target') . DS . sprintf('%s.%s', $result, 'css');
         $this->assertFileExists($file);
@@ -226,7 +212,7 @@ class AssetsCreatorTest extends TestCase
 
         //Tests array
         $result = (new AssetsCreator(['test', 'test2'], 'css'))->create();
-        $this->assertRegExp('/^[a-z0-9]+$/', $result);
+        $this->assertRegExp('/^[\w\d]+$/', $result);
 
         $file = Configure::read(ASSETS . '.target') . DS . sprintf('%s.%s', $result, 'css');
         $this->assertFileExists($file);
@@ -243,7 +229,7 @@ class AssetsCreatorTest extends TestCase
     public function testCreateWithJs()
     {
         $result = (new AssetsCreator('test', 'js'))->create();
-        $this->assertRegExp('/^[a-z0-9]+$/', $result);
+        $this->assertRegExp('/^[\w\d]+$/', $result);
 
         $file = Configure::read(ASSETS . '.target') . DS . sprintf('%s.%s', $result, 'js');
         $this->assertFileExists($file);
@@ -254,7 +240,7 @@ class AssetsCreatorTest extends TestCase
 
         //Tests array
         $result = (new AssetsCreator(['test', 'test2'], 'js'))->create();
-        $this->assertRegExp('/^[a-z0-9]+$/', $result);
+        $this->assertRegExp('/^[\w\d]+$/', $result);
 
         $file = Configure::read(ASSETS . '.target') . DS . sprintf('%s.%s', $result, 'js');
         $this->assertFileExists($file);
@@ -314,7 +300,7 @@ class AssetsCreatorTest extends TestCase
     public function testFilename()
     {
         $asset = new AssetsCreator('test', 'css');
-        $this->assertRegExp('/^[a-z0-9]+$/', $asset->filename());
+        $this->assertRegExp('/^[\w\d]+$/', $asset->filename());
     }
 
     /**