+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     extends `Exception`;
 * replaced `InternalErrorException` with `InvalidArgumentException` and 
     `RuntimeException`. This allows compatibility with CakePHP 3.6 branch;
+* added `IntegrationTestCase` and `TestCase` classes;
 * now it uses the `mirko-pagliai/php-tools` package. This also replaces
     `mirko-pagliai/reflection`.